Created policy view component.
